Introduce automake to the lot. The Makefile.in have been removed, and Makefile.am are now the preferred (and supported) way of changing makefiles. Use libtool.
These changes allow to build the examples at the same time as the library, and reduce
the amount of Makefile glue that needs to be written for a new example.
Basically, one needs:
1. entry in configure.in AC_OUTPUT to generate the Makefile
2. entry in the parent directories' Makefile.am SUBDIR line.
3. the Makefile.am

Some third-party code in this commit:
config/libtool15.m4 is from libtool. This file should be installed on every machine that
                    has libtool, but I prefer having a version in use I control.
config/mdl_have_opengl.m4 is a comprehensive check for opengl libraries.
config/pthread.m4 checks for the name of the pthread library. This was inspired
                  by the similar check in gnome.