Generate documentation from source code
Most updates of the documentation could automatically be done when we extract
the informations out of the source code. For this we need to insert the
detailed description of a function into a documentation block before the
function and let another tool extract them. For this purpose doxygen provided
a output plugin for xml which extracts as much information as possible for all
files and outputs all related informations in specific files. So we have
a s3d(w)_8h.xml with all informations to the public api functions in a
doxygen scheme. Another tool can now parse this xml data and transform it into
a partial docbook file with all functions.

Steps to be taken when generating an updated documentation are:
- $ doxygen xml.doxygen
- $ ./extractdefinitions.py
- running docbook2html as part of the normal build process

Currently only doxygen 1.5.6 and python 2.5.2 were tested and so these steps
should not be part of the normal build process.

To have the descriptions near to code should be lowering the barriers to update
them and results in a better overall documentation quality.