Read incoming data from aligned address
When mips reads data from a unaligned address a sigbus will be generated.If it
we will read it into a general register the kernel will emulate the read after
noticing the cpu exception, but if we read into a fpu register the program will
receive a sighub and stops. As workaround we read the 8 bit cmd and the 16 bit
length into the input buffer and the user data will be saved to a address
4 bytes or more away from the beginning of the start of te input buffer.
This should be an aligned address. To establish this we use int_least32_t as
reference data type.