Change docbook compiler to xmlto
We are using a xml based version of docbook and not the (original) sgml
version. (Open)Jade will fail on most distributions when it tries to compile
the xml docbook. The correct way to build them is to use xsl to transform the
document into the target file.
The usage of raw xsltproc depends too much on the environment and will
definitely fail on some distributions. The easiest way to get the same
functionality as before is to use xmlto - a bash script which will be offered
by the distributions and should always know where the correct xsl files can
be found.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>