Mark functions with important return as warn_unused_return
Functions which will create dangling resources when nothing is done with the
return creates a warning when the return is not saved inside the calling
function.
This could for example also added to functions which create a security risk or
where it is a general bug to omit the return.