Mark functions without side effect and local storage as const
Functions which doesn't read from global memory and which will not write to
global memory can be declared as const. The function must not read anything but
the function parameters to create the return value.
This makes it safe for the optimizer in some cases to create less calls then
specified by the source code.