server: Process tcp data even if shm is deactivated
If server is compiled with `cmake . -DHAVE_SHM=0` no data of a process
will be consumed. If it is compiled with SHM support turn will be reset
to 1 in each shm_main. To get the same behaviour without shm_main we set
turn to 1 in each call to network_main.