server: Include glext.h for post-1.2 OpenGL constants
To compile s3d on cygwin with windows we need a different name mangling
than the one used by mesa's opengl headers. We should include the one
provided by w32api but the GL/gl.h doesn't provide constants for
extension introduced later then OpenGL 1.2. The best way to get it
working with mesa and w32api is to include GL/glext.h even if we don't
need it on systems using mesa.