Remove binary files without free copyright status
All files from Martin Stern aren't under a (known) free license. This
could end in s3d being marked as non-free software in distributions with
an hard copyright check. Non of the files are used by s3d or related
programs and should not create any missing dependencies when they get
removed.
This commit can be reverted when the copyright status changes to a
compatible license.