Correct check for new style gpsd structs in s3dosm
gpsd changed the name of the defines which were used to detect if it is
a version with their new style data structures. It is a better idea to
check if the the structure gps_data_t has a member with the name fix
instead of guessing it by looking at some defines which maybe changes
over time.