Only strip out libs3d options from argv
The idea of parse_args of libs3d was to parse all known opts and than
stop before the parameter for the program starts. argv and argc is than
modified to start at the first option not recognised by libs3d.
This implementation was faulty and stripped of all valid options
beginning with a dash in different versions of getopt. A better idea is
to parse all parameters regardless of their first character and move
them to the left to override libs3d options. The order in which libs3d
and application specific option occurred isn't important anymore.