Make the dependency to libg3d 0.0.7 more visible
Their are currently users which see that cmake cannot find libg3d even
if they have libg3d-0.0.6 installed and then overriding the INCLUDE
and LIBRARY informations for libg3d in their cmake configuration.
It is better to make the information about the version dependency more
visible for users during an configure failure.