Add test for correct linking in maintainers guide
It is important that all libraries have no unresolved symbols or
otherwise the executable has to handle these symbols. This would add
extra libraries to these executables without that they uses them. In
most cases this would create unusable executables when binaries and
libraries are getting updated independently when new dependencies are
getting introduced. This can be tested with the option --no-undefined of
GNU ld.

The other situation is that an executable doesn't specify all libraries
which must be loaded to resolve all symbols it uses, but a library would
also use the necessary second library. In this case the linker would
look at the DT_NEEDED entries of the first library, find the second
library and tries to resolve the symbols in the executable using this
indirect linked library. This could break the executable when the first
library removes its dependency to the second library.

The standard behavior of GNU ld is to check entries in DT_NEEDED. This
can be disabled with --no-add-needed. GNU gold doesn't do it in the
moment and doesn't provide the option either.