Depend on libgps 2.90 for s3dosm GPS support
gpsd as stated with version 2.90 that the binary breakage with nearly
each update should stop and that they want to have a better binary
stability by now.

As this is the first hard break in the development of it, we start to
support only the current version and all following compatible versions
instead of fixing each small breakage using cmake checks due to the old
development model with each minor update of gpsd.