Allow whole-program optimization when using ENABLE_FINAL
gcc can only optimize the whole program when it knows that the current
compilation unit is really the whole program. To support shared objects
as whole-program compilation units we must mark exported functions as
externally_visible - which means something like "definitelly not
static". This is complete independent to the "visibility=hidden"
behavior we try to use for shared objects.

It doesn't enable -fwhole-program by default when using ENABLE_FINAL due
to the gcc bug #40546.