libs3d: Don't accidently used INFINITY from math.h
math.h has a constant defined under the name INFINITY which is only
defined for floats and not for double. As result the triangulation of
some segments failed from time to time.

This happens because we also include math.h and it may override our
definition of INFINITY. To solve this we just can use HUGE_VAL from
math.h which  is defined as the largest possible positive double.