s3dvt: Reimplement PTY/TTY mode using glibc
s3dvt used to open a pty/pts pair by searching /dev for those files.
Those paths have been changed and the overall procedure to get a slave
pseudo terminal is now dramatically different.

Glibc provides enough functionality to create a pair of pseudo terminal
master and slave without doing a lot of filesystem work. It is possible
that on non linux system extra functions must be implemented to get it
working. Those functions are getpt which should open /dev/ptmx and
ptsname_r which stores the name of the slave device in an extra buffer.