s3dvt: Fix FTBFS with glibc 2.30
The s3d code used stropts.h and thought it uses `isastream`, but that
was wrong.
One might think [1] this is a valid call to check if something is a stream.
But in reality this was never implemented [2].
In glibc 2.30 this now got [3]:
 "The obsolete and never-implemented XSI STREAMS header files <stropts.h>
 and <sys/stropts.h> have been removed."
Since it always returned -1 it is equivalent in function to just drop
the related lines.

[1]: https://linux.die.net/man/3/isastream
[2]: https://linux.die.net/man/2/isastream
[3]: https://savannah.gnu.org/forum/forum.php?forum_id=9515