libs3d: Fix rendering of letter with 0 area
The definition of an hole in a freetype font was not clear in the
triangulation code. It was either:

* area < 0
* area <= 0

The definition of a contour was unclear in the same way. This then caused
rejections of letters with 0 areas because none of the processing code
paths was responsible for it. Only the rejection path for the whole letter
was triggered. A clear definition avoids this:

* area < 0: hole
* area >= 0: contour

Signed-off-by: Sven Eckelmann <sven@narfation.org>